+++ bluetape4k/core/src/main/kotlin/io/bluetape4k/concurrent/virtualthread/VirtualFutureExtensions.kt
@@ -2,6 +2,7 @@ package io.bluetape4k.concurrent.virtualthread
 
 import io.bluetape4k.concurrent.asCompletableFuture
 import io.bluetape4k.concurrent.sequence
+import io.bluetape4k.concurrent.virtualthread.api.VirtualThreads
 import io.bluetape4k.utils.ShutdownQueue
 import java.time.Duration
 import java.util.concurrent.CompletableFuture