+++ virtualthread/api/src/test/kotlin/io/bluetape4k/concurrent/virtualthread/api/VirtualThreadsTest.kt
@@ -1,4 +1,4 @@
-package io.bluetape4k.concurrent.virtualthread
+package io.bluetape4k.concurrent.virtualthread.api
 
 import io.bluetape4k.assertions.shouldBeEqualTo
 import io.bluetape4k.assertions.shouldBeTrue