+++ virtualthread/jdk21/src/test/kotlin/io/bluetape4k/concurrent/virtualthread/jdk21/Jdk21TaskContextTest.kt
@@ -1,6 +1,6 @@
 package io.bluetape4k.concurrent.virtualthread.jdk21
 
-import io.bluetape4k.concurrent.virtualthread.TaskContext
+import io.bluetape4k.concurrent.virtualthread.api.TaskContext
 import io.bluetape4k.logging.KLogging
 import io.bluetape4k.logging.debug
 import io.bluetape4k.assertions.shouldBeEqualTo